# [FIRRTL][Dedup] Alter dedup group handling, avoid exponential behavior.