@youngar 2024-05-03
Just to connect the dots, this is similar to what we do to prevent memories in the testharness and (i think) prefix-module-groups from deduping with each other: https://github.com/llvm/circt/pull/2968